refactor: :recycle: Move DNS check from sync_complete to inside diff.
In order to have diffs be more accurate I've moved the DNS query function that was done after the sync was complete to instead be done in the D42 adapter while data is being loaded. This should theoretically allow the IPAddrs and Interfaces that were created by that process to be in the diff and no longer be constantly added/deleted.

Move DNS function inside diff adapter.